+++ src/app/components/index.ts
@@ -1,5 +1,0 @@
-export * from './first/first.component';
-export * from './product/product.component'
-export * from './product-list/product-list.component'
-export * from './cart-list/cart-list.component'
-export * from './cart-item/cart-item.component'