+++ src/app/shared/category.ts
@@ -1,6 +1,6 @@
 export enum ECategory {
-  mensClothing ="Men's clothing",
+  mensClothing = "Men's clothing",
   jewelery = "Jewelery",
   electronics = "Electronics",
-  womensClothing ="Women's clothing"
+  womensClothing = "Women's clothing"
 }