+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Task 1
 
-### Created
+### Created:
 - FirstComponent
 - ProductComponent
 - ProductListComponent
@@ -10,3 +10,22 @@
 - Product moc
 - Product type
 - CartItemComponent (empty)
+
+
+# Task 2
+
+### Created:
+- CartModule,
+- ProductsModule,
+- SharedModule
+- zoom directive
+- category pipe
+
+### Implemented:
+- adding products
+- purchase amount
+- product removal
+- opening closing the cart
+- quantity and amount of product
+- zoom the cart icon on the hover
+